GenRFS

Support for random finite sets in Gen.jl

Getting Started

Installation

Example

Reference

Random Finite Sets

Random Finite Elements

API

GenRFS.AbstractRFS — Type

Random Finite Set

GenRFS.AssociationRecord — Type

Contains a record of the top n data associations

Has two fields:

table

logscores

GenRFS.RFSElements — Type

A type alias for random finite elements

GenRFS.associations — Method

Computes the logscore of every correspondence

Returns a vector where each element is indexed in the partition table.

GenRFS.bounds — Method

The lower and upper bounds of map cardinality

GenRFS.cardinality — Method

The log probability P(|{x}| | rfe) The log probability of the cardinality of an observation for an RFE.

GenRFS.contains — Method

Whether the given RFS can support the cardinality of the observation

GenRFS.partition — Method

Generates the partition table for a given set of size n.

Only valid when the random finite set contains the observed set.

GenRFS.partition_press — Method

Returns the combination table for correspondence cardinality.

For an ordered set of morphological bounds, returns a table where each row describes a different combination of cardinalities of assignments.

ie

julia> GenRFS.partition_press([4,1,1],[0,0,0], 4)
4-element Array{Array{Int64,1},1}:
 [2, 1, 1]
 [3, 1, 0]
 [3, 0, 1]
 [4, 0, 0]

GenRFS.partition_push — Method

Returns a table of all associations for a given cardinality combination.

Rows are the complete disjoint partitioning of the elements in xs onto elements described in cs

eg.

julia> GenRFS.partition_push([2,1,1], [1,2,3,4])
12-element Array{Array{Array{Int64,1},1},1}:
 [[1, 2], [3], [4]]
 [[1, 2], [4], [3]]
 [[1, 3], [2], [4]]
 [[1, 3], [4], [2]]
 [[1, 4], [2], [3]]
 [[1, 4], [3], [2]]
 [[2, 3], [1], [4]]
 [[2, 3], [4], [1]]
 [[2, 4], [1], [3]]
 [[2, 4], [3], [1]]
 [[3, 4], [1], [2]]
 [[3, 4], [2], [1]]

GenRFS.partition_table — Method

Returns the partition table given the morphological bounds of RFEs

Each top level entry is a possible partition. Within each partions, each entry describes the associations of observation elements to a given element.

The elements are ordered by upper morphological bound in descending order.

NOTE: Isomorphic elements are not currently supported

GenRFS.support — Method

The probability of P(x | dist(rfe)) The loglikelihood of a component, x, given the distribution contained by the element.

Index

GenRFS.AbstractRFS

GenRFS.AssociationRecord

GenRFS.RFSElements

GenRFS.associations

GenRFS.bounds

GenRFS.cardinality

GenRFS.contains

GenRFS.partition

GenRFS.partition_press

GenRFS.partition_push

GenRFS.partition_table

GenRFS.support